==> 1_preparation.ipynb <==
%pip install git+https://github.com/AI4Finance-LLC/FinRL-Library.git
%pip install yfinance
%pip install pandas
%pip install numpy
%pip install matplotlib
%pip install stockstats
%pip install gym
%pip install stable-baselines3[extra]
%pip install tensorflow
%pip install ta
%pip install PyPortfolioOpt
%pip install akshare
%pip install git+https://github.com/quantopian/pyfolio

from config import config
from datetime import datetime
import os

current_time = datetime.now().strftime("%y%m%d_%H%M%S")

data_dir = config.DATA_SAVE_DIR + "/" + current_time
trained_model_dir = config.TRAINED_MODEL_DIR + "/" + current_time
tensorboard_log_dir = config.TENSORBOARD_LOG_DIR + "/" + current_time
result_dir = config.RESULTS_DIR + "/" + current_time

os.makedirs(data_dir)
os.makedirs(trained_model_dir)
os.makedirs(tensorboard_log_dir)
os.makedirs(result_dir)

%store data_dir
%store trained_model_dir
%store tensorboard_log_dir
%store result_dir

current_time

import akshare as ak

%matplotlib inline

fund_name_em_df = ak.fund_name_em()
fund_name_em_df.to_csv(f"{config.DATA_SAVE_DIR}/fund_name.csv", index=False)
fund_name_em_df

import utils

fund_open_fund_rank_em_df = utils.fund_open_fund_rank_em()
fund_open_fund_rank_em_df.to_csv(f"{config.DATA_SAVE_DIR}/fund_rank.csv", index=False)
fund_open_fund_rank_em_df
