==> 2_data.ipynb <==
import pandas as pd
from config import config


fund_open_fund_rank_em_df = pd.read_csv(
    f"{config.DATA_SAVE_DIR}/fund_rank.csv", dtype=str
)


fund_open_fund_rank_em_df.set_index("序号", inplace=True)
fund_open_fund_rank_em_df

fund_filtered = fund_open_fund_rank_em_df[
    fund_open_fund_rank_em_df["成立日期"] <= config.THRESH_DATE
].iloc[:10]
fund_filtered

fund_choose = fund_filtered["基金代码"].tolist()

%store fund_choose

fund_filtered.loc[:, ["基金代码", "基金简称"]]

import akshare as ak
import os

fund_choose_data = pd.DataFrame()
data = pd.DataFrame()
data_dir = config.DATA_SAVE_DIR
nav_dir = f"{data_dir}/NAV"
%store -r data_dir

for fund in fund_choose:
    if not os.path.exists(f"{nav_dir}/{fund}.csv"):
        fund_open_fund_info_em_df = ak.fund_open_fund_info_em(symbol=fund)
        if fund_open_fund_info_em_df.empty:
            print(f"fund {fund} is empty")
            continue
        else:
            fund_open_fund_info_em_df.to_csv(f"{nav_dir}/{fund}.csv", index=False)
    else:
        fund_open_fund_info_em_df = pd.read_csv(f"{nav_dir}/{fund}.csv")
    series = fund_open_fund_info_em_df.set_index("净值日期")["单位净值"]
    fund_choose_data[fund] = series
    fund_open_fund_info_em_df.insert(1, "基金代码", fund)
    data = pd.concat([data, fund_open_fund_info_em_df], axis=0)

fund_choose_data = fund_choose_data.dropna(how="any", axis=0)
fund_choose_data = fund_choose_data.reset_index().rename(columns={"净值日期": "date"})
fund_choose_data = fund_choose_data[fund_choose_data["date"] < config.END_DATE]
fund_choose_data.to_csv(f"{data_dir}/fund_choose_data.csv", index=False)
data = (
    data[data["净值日期"].isin(fund_choose_data["date"])]
    .sort_values(by=["净值日期"])
    .reset_index(drop=True)
)
%store data
data

fund_choose_data.describe().T
