==> 3_feature.ipynb <==
import pandas as pd

data = pd.DataFrame()
%store -r data
data

data = data.loc[:, ["净值日期", "基金代码", "单位净值"]]
data.columns = ["date", "fund", "nav"]
data

from config import config

data = data.sort_values(["date", "fund"], ignore_index=True)



data.index = data.date.factorize()[0]



cov_list = []  # create empty list for storing coveriance matrices at each time step



for i in range(config.APPROX_BDAYS_PER_YEAR, len(data.index.unique())):


    data_lookback = data.loc[i - config.APPROX_BDAYS_PER_YEAR : i, :]


    price_lookback = data_lookback.pivot_table(
        index="date", columns="fund", values="nav"
    )


    return_lookback = price_lookback.dropna().pct_change()


    covs = return_lookback.cov().values


    covs = covs / covs.max()
    cov_list.append(covs)



df_cov = pd.DataFrame({"date": data.date.unique()[config.APPROX_BDAYS_PER_YEAR :], "cov": cov_list})



data_with_cov = data.merge(df_cov, on="date")


data_with_cov = data_with_cov.sort_values(["date", "fund"]).reset_index(drop=True)
data_with_cov = data_with_cov[data_with_cov["date"] >= config.START_DATE]



data_with_cov

data_dir = ""
%store -r data_dir

print("Training Data: ", "from ", config.START_DATE, " to ", config.TRAIN_DATE)
print("Testing Data: ", "from ", config.TRAIN_DATE, " to ", config.END_DATE)
choose_data = pd.read_csv(f"{data_dir}/fund_choose_data.csv")

train_data = data_with_cov[data_with_cov["date"] < config.TRAIN_DATE]
test_data = data_with_cov[data_with_cov["date"] >= config.TRAIN_DATE]
train_choose = choose_data[choose_data["date"] < config.TRAIN_DATE]
test_choose = choose_data[choose_data["date"] >= config.TRAIN_DATE]

train_data.index = train_data["date"].factorize()[0]
test_data.index = test_data["date"].factorize()[0]
train_choose = train_choose.reset_index(drop=True).set_index(["date"])
test_choose = test_choose.reset_index(drop=True).set_index(["date"])

%store train_data
%store test_data
%store train_choose
%store test_choose

test_choose
