==> 5_train.ipynb <==
import pandas as pd

train_data = pd.DataFrame()
test_data = pd.DataFrame()
result_dir=""

%store -r train_data
%store -r test_data
%store -r result_dir

train_data

fund_dimension = len(train_data["fund"].unique())
state_space = fund_dimension
print(f"Fund Dimension: {fund_dimension}, State Space: {state_space}")

from env_portfolio import PortfolioEnv
from models import DRLAgent
from config import config


weights_initial = [1 / fund_dimension] * fund_dimension


env_kwargs = {
    "hmax": 100000,
    "initial_amount": 1000000,
    "transaction_cost_pct": 0.001,
    "state_space": state_space,
    "fund_dim": fund_dimension,
    "tech_indicator_list": [],
    "action_space": fund_dimension,
    "reward_scaling": 0,
    "initial_weights": [1 / fund_dimension] * fund_dimension,
    "turbulence_threshold": config.RISK_FREE_RATE,
    "lookback": config.APPROX_BDAYS_PER_YEAR,
    "result_dir": result_dir,
}


e_train_gym = PortfolioEnv(df=train_data, **env_kwargs)


env_train, _ = e_train_gym.get_sb_env()


agent = DRLAgent(env=env_train)

A2C_PARAMS = {"n_steps": 5, "ent_coef": 0.005, "learning_rate": 0.0002}
model_a2c = agent.get_model(model_name="a2c", model_kwargs=A2C_PARAMS)
trained_a2c = agent.train_model(
    model=model_a2c, tb_log_name="a2c", total_timesteps=50000
)

model_td3 = agent.get_model(model_name="td3")
trained_td3 = agent.train_model(
    model=model_td3, tb_log_name="td3", total_timesteps=50000
)

PPO_PARAMS = {
    "n_steps": 2048,
    "ent_coef": 0.005,
    "learning_rate": 0.0001,
    "batch_size": 128,
}
model_ppo = agent.get_model("ppo", model_kwargs=PPO_PARAMS)
trained_ppo = agent.train_model(
    model=model_ppo, tb_log_name="ppo", total_timesteps=50000
)

DDPG_PARAMS = {"batch_size": 128, "buffer_size": 50000, "learning_rate": 0.001}
model_ddpg = agent.get_model("ddpg", model_kwargs=DDPG_PARAMS)
trained_ddpg = agent.train_model(
    model=model_ddpg, tb_log_name="ddpg", total_timesteps=50000
)

model_sac = agent.get_model("sac")
trained_sac = agent.train_model(
    model=model_ddpg, tb_log_name="sac", total_timesteps=50000
)

e_trade_gym = PortfolioEnv(df=train_data, **env_kwargs)
env_trade, obs_trade = e_trade_gym.get_sb_env()

a2c_train_daily_return, a2c_train_weights = DRLAgent.DRL_prediction(
    model=trained_a2c, test_data=train_data, test_env=env_trade, test_obs=obs_trade
)

td3_train_daily_return, td3_train_weights = DRLAgent.DRL_prediction(
    model=trained_td3, test_data=train_data, test_env=env_trade, test_obs=obs_trade
)

ppo_train_daily_return, ppo_train_weights = DRLAgent.DRL_prediction(
    model=trained_ppo, test_data=train_data, test_env=env_trade, test_obs=obs_trade
)

ddpg_train_daily_return, ddpg_train_weights = DRLAgent.DRL_prediction(
    model=trained_ddpg, test_data=train_data, test_env=env_trade, test_obs=obs_trade
)

sac_train_daily_return, sac_train_weights = DRLAgent.DRL_prediction(
    model=trained_sac, test_data=train_data, test_env=env_trade, test_obs=obs_trade
)

%store a2c_train_daily_return
%store td3_train_daily_return
%store ppo_train_daily_return
%store ddpg_train_daily_return
%store sac_train_daily_return

e_trade_gym = PortfolioEnv(df=test_data, **env_kwargs)
env_trade, obs_trade = e_trade_gym.get_sb_env()
result_dir = ""
%store -r result_dir

a2c_test_daily_return, a2c_test_weights = DRLAgent.DRL_prediction(
    model=trained_a2c, test_data=test_data, test_env=env_trade, test_obs=obs_trade
)
a2c_test_weights.to_csv(f"{result_dir}/a2c_test_weights.csv")
a2c_test_daily_return

td3_test_daily_return, td3_test_weights = DRLAgent.DRL_prediction(
    model=trained_td3, test_data=test_data, test_env=env_trade, test_obs=obs_trade
)
td3_test_weights.to_csv(f"{result_dir}/td3_test_weights.csv")
td3_test_daily_return

ppo_test_daily_return, ppo_test_weights = DRLAgent.DRL_prediction(
    model=trained_ppo, test_data=test_data, test_env=env_trade, test_obs=obs_trade
)
ppo_test_weights.to_csv(f"{result_dir}/ppo_test_weights.csv")
ppo_test_daily_return

ddpg_test_daily_return, ddpg_test_weights = DRLAgent.DRL_prediction(
    model=trained_ddpg, test_data=test_data, test_env=env_trade, test_obs=obs_trade
)
ddpg_test_weights.to_csv(f"{result_dir}/ddpg_test_weights.csv")
ddpg_test_daily_return

sac_test_daily_return, sac_test_weights = DRLAgent.DRL_prediction(
    model=trained_sac, test_data=test_data, test_env=env_trade, test_obs=obs_trade
)
sac_test_weights.to_csv(f"{result_dir}/sac_test_weights.csv")
sac_test_daily_return

a2c_test_portfolio = a2c_test_daily_return.copy()
a2c_test_returns = a2c_test_daily_return.copy()

td3_test_portfolio = td3_test_daily_return.copy()
td3_test_returns = td3_test_daily_return.copy()

ppo_test_portfolio = ppo_test_daily_return.copy()
ppo_test_returns = ppo_test_daily_return.copy()

ddpg_test_portfolio = ddpg_test_daily_return.copy()
ddpg_test_returns = ddpg_test_daily_return.copy()

sac_test_portfolio = sac_test_daily_return.copy()
sac_test_returns = sac_test_daily_return.copy()

%store a2c_test_portfolio
%store a2c_test_returns 

%store td3_test_portfolio
%store td3_test_returns 

%store ppo_test_portfolio
%store ppo_test_returns 

%store ddpg_test_portfolio
%store ddpg_test_returns 

%store sac_test_portfolio
%store sac_test_returns
